Remove beaker to fully utilize litmus and litmus:acceptance:parallel
