+++ .puppet-lint.rc
@@ -1,0 +1,1 @@
+--relative
