+++ metadata.json
@@ -40,6 +40,6 @@
     "dsc"
   ],
   "pdk-version": "1.10.0",
-  "template-url": "https://github.com/puppetlabs/pdk-templates.git#1.10.0",
-  "template-ref": "1.10.0-0-gbba9ac3"
+  "template-url": "https://github.com/puppetlabs/pdk-templates#master",
+  "template-ref": "heads/master-0-g7885b5e"
 }
