+++ provision.yaml
@@ -1,6 +1,6 @@
 default:
   provisioner: vmpooler
-  images: ['win-2012r2-x86_64']
+  images: ['win-2012r2-wmf5-x86_64']
 release_checks:
   provisioner: vmpooler
-  images: ['win-2012-x86_64', 'win-2012r2-x86_64', 'win-2016-x86_64', 'win-2019-x86_64']
+  images: ['win-2012r2-wmf5-x86_64', 'win-2016-core-x86_64', 'win-2019-core-x86_64']
