+++ spec/acceptance/basic_functionality/puppet_apply_noop_dsc_manifest_spec.rb
@@ -1,28 +1,34 @@
 require 'spec_helper_acceptance'
 
 describe 'Apply DSC Resource Manifest in "noop" Mode Using "puppet apply"' do
-  test_dir_name = 'test'
-  test_dir_path = "C:/#{test_dir_name}"
-  test_file_path = "#{test_dir_path}/test.txt"
-  test_file_contents = 'catcat'
-
-  dsc_manifest = test_file_path_manifest(test_dir_path, test_file_path, test_file_contents)
-
   context 'Apply noop manifest and verify no files created' do
-    windows_agents.each do |agent|
-      it 'runs noop on manifest does not create files after noop manifest application' do
-        execute_manifest_on(agent, dsc_manifest, :noop => true, :catch_failures => true)
-
-        expect_failure('Expect failure because nothing should have changed') do
-          assert_dsc_resource(
-              agent,
-              'File',
-              'PSDesiredStateConfiguration',
-              :DestinationPath => test_file_path,
-              :Contents => test_file_contents
-          )
-        end
-      end
+    it 'runs noop on manifest does not create files after noop manifest application' do
+      dsc_manifest = <<-Manifest
+      dsc_file {'tmp_folder':
+          dsc_ensure          => 'present',
+          dsc_type            => 'Directory',
+          dsc_destinationpath => "C:/test",
+      }
+    
+      dsc_file {'tmp_file':
+          dsc_ensure          => 'present',
+          dsc_type            => 'File',
+          dsc_destinationpath => "C:/test/test.txt",
+          dsc_contents        => 'catcat'
+      }
+      Manifest
+      apply_manifest(dsc_manifest, :noop => true)
+      
+      result = assert_dsc_resource(
+            'File',
+            'PSDesiredStateConfiguration',
+            true, # Passing true to expect_ps_failures. We expect this to fail.
+            true, # Passing true to expect_dsdc_failures. We expect dsc to fail to find the resource.
+            :DestinationPath => "C:/test/test.txt",
+            :Contents => 'catcat',
+        )
+      result_string = strip_crln(result.stdout)
+      expect(result_string).to match(/cannot find the path specified/)
     end
   end
 end
