--- spec/acceptance/nodesets/centos-7-x64.yml
@@ -1,10 +1,0 @@
-HOSTS:
-  centos-7-x64:
-    roles:
-      - agent
-      - default
-    platform: el-7-x86_64
-    hypervisor: vagrant
-    box: puppetlabs/centos-7.2-64-nocm
-CONFIG:
-  type: foss
