--- spec/acceptance/nodesets/debian-8-x64.yml
@@ -1,10 +1,0 @@
-HOSTS:
-  debian-8-x64:
-    roles:
-      - agent
-      - default
-    platform: debian-8-amd64
-    hypervisor: vagrant
-    box: puppetlabs/debian-8.2-64-nocm
-CONFIG:
-  type: foss
