--- spec/acceptance/nodesets/docker/centos-7.yml
@@ -1,12 +1,0 @@
-HOSTS:
-  centos-7-x64:
-    platform: el-7-x86_64
-    hypervisor: docker
-    image: centos:7
-    docker_preserve_image: true
-    docker_cmd: '["/usr/sbin/init"]'
-    # install various tools required to get the image up to usable levels
-    docker_image_commands:
-      - 'yum install -y crontabs tar wget openssl sysvinit-tools iproute which initscripts'
-CONFIG:
-  trace_limit: 200
