--- spec/acceptance/nodesets/docker/debian-8.yml
@@ -1,11 +1,0 @@
-HOSTS:
-  debian-8-x64:
-    platform: debian-8-amd64
-    hypervisor: docker
-    image: debian:8
-    docker_preserve_image: true
-    docker_cmd: '["/sbin/init"]'
-    docker_image_commands:
-      - 'apt-get update && apt-get install -y net-tools wget locales strace lsof && echo "en_US.UTF-8 UTF-8" > /etc/locale.gen && locale-gen'
-CONFIG:
-  trace_limit: 200
